Remove SDL.h from our headers

+++ src/utils/enum-array.h
@@ -7,31 +7,29 @@
 namespace EnumArray
 {
 
-
 template <typename t_Index, typename t_Element>
 struct EnumArray
 {
   using IndexType = t_Index;
   using ElementType = t_Element;
-  static size_t const size = (size_t)IndexType::COUNT - (size_t)IndexType::START;
+  static u32 const size = (u32)IndexType::COUNT - (u32)IndexType::START;
 
   ElementType const&
   operator[](IndexType index) const
   {
     assert(index >= IndexType::START && index < IndexType::COUNT);
-    size_t offset = (size_t)index - (size_t)IndexType::START;
+    u32 offset = (u32)index - (u32)IndexType::START;
     return members[offset];
   }
   ElementType&
   operator[](IndexType index)
   {
     assert(index >= IndexType::START && index < IndexType::COUNT);
-    size_t offset = (size_t)index - (size_t)IndexType::START;
+    u32 offset = (u32)index - (u32)IndexType::START;
     return members[offset];
   }
 
   ElementType members[size];
 };
-
 
 }  // namespace EnumArray
